Rename `systemHandler` package to `system` and update references across codebase.
