+++ internal/system/system_test.go
@@ -1,4 +1,4 @@
-package systemHandler
+package system
 
 import (
 	"net/http"
