+++ router/router.go
@@ -37,7 +37,7 @@ func SetupRouter(dbConn db.DBTX) http.Handler {
 }
 
 func registerSystemCallHandler(r *chi.Mux) {
-	h := systemHandler.NewSystemCallHandler()
+	h := system.NewSystemCallHandler()
 	r.Get("/api/time", h.GetTime)
 	r.Get("/api/version", h.GetVersion)
 }
